requirements.txt
pdfplumber
pdf2image
pytesseract
python-docx
Pillow